cleanup code for Go 1.5
